Refactor Travis rig so devs can run tests locally
When using this template I moved the test command into a rake task, which makes it more convenient to run tests locally. In particular I had a bug which Travis uncovered – a legitimate memory bug – which manifested consistently in xcodebuild but not in Xcode itself.  Autoreleasing behaves a little bit differently between the two.

The point is, it's useful for a developer to easily run the tests locally.

It's also easier to bring new contributors into a project when the dev installation is scripted.

This change adds a layer of indirection through rake. I'm not normally a fan of indirection, but I do think developers should be able to run the tests locally. A shell script would work here, but in my experience projects quickly outgrow them and they are harder to read.